requirements.txt
torch>=2.1.0
torchvision>=0.16.0
timm>=0.9.12
albumentations>=1.3.1
opencv-python-headless>=4.8.0
Pillow>=10.0.0
numpy>=1.24.0
pandas>=2.0.0
mlflow>=2.9.0
pyyaml>=6.0
tqdm>=4.66.0
scikit-learn>=1.3.0
grad-cam>=1.4.8
streamlit>=1.30.0
plotly>=5.18.0
matplotlib>=3.7.0
dvc>=3.30.0
pytest>=7.4.0
pytest-cov>=4.1.0
flake8>=6.1.0